Fix to disambiguate which String constructor to call.  Apparently the gcc on the s390 needs a little coaxing.

+++ test/c++providers/indication/OW_IndicationProviderTest1.cpp
@@ -223,7 +223,7 @@ private:
 			iToAdd.setProperty("SystemCreationClassName", CIMValue("CIM_System"));
 			iToAdd.setProperty("SystemName", CIMValue("localhost"));
 			iToAdd.setProperty("CreationClassName", CIMValue("OW_IndicationProviderTest1"));
-			iToAdd.setProperty("DeviceID", CIMValue(String(m_insts.size())));
+			iToAdd.setProperty("DeviceID", CIMValue(String(UInt32(m_insts.size()))));
 			// PowerOnHours is our property that will be modified
 			iToAdd.setProperty("PowerOnHours", CIMValue(UInt64(m_insts.size())));
 			m_insts.push_back(iToAdd);
