+++ test/c++providers/indication/OW_IndicationProviderTest2.cpp
@@ -481,7 +481,7 @@ public:
 			iToAdd.setProperty("SystemCreationClassName", CIMValue("CIM_System"));
 			iToAdd.setProperty("SystemName", CIMValue("localhost"));
 			iToAdd.setProperty("CreationClassName", CIMValue("OW_IndicationProviderTest2"));
-			iToAdd.setProperty("DeviceID", CIMValue(String(m_insts.size())));
+			iToAdd.setProperty("DeviceID", CIMValue(String(UInt32(m_insts.size()))));
 			// PowerOnHours is our property that will be modified
 			iToAdd.setProperty("PowerOnHours", CIMValue(UInt64(m_insts.size())));
 			if (creat > 0)
